pyproject.toml
[project]
name = "acv-final-project"
version = "0.1.0"
description = "ASL video-to-text translation using Qwen2.5-VL-2B with Unsloth + QLoRA"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.6",
    "diffusers>=0.37.1",
    "peft>=0.15.0",
    "trl>=0.18.2,<=0.24.0",
    "unsloth>=2025.7.2",
    "unsloth-zoo>=2026.4",
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "xformers>=0.0.29",
    "datasets>=4.3.0",
    "opencv-python-headless>=4.13.0.92",
    "numpy>=2.4.4",
    "pandas>=2.2.0",
    "pillow>=11.0.0",
    "sacrebleu>=2.6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
